Comment out the unused import scss file icon
