+++ modules/_metronic/assets/sass/components/_variables.custom.scss
@@ -6,9 +6,9 @@
 
 // Theme colors
 // Primary
-$primary:                   #0076b9; // #04c8c8
-$primary-active:            #0b5ed7; // #0a58ca | metronic = #00afaf
-$primary-light:             #dcfdfd;
+$primary:                   #0076b9; // metronic = #04c8c8
+$primary-active:            #225098; // #0b5ed7 | #0a58ca | metronic = #00afaf
+$primary-light:             #d6ecfb; // metronic = #dcfdfd
 $primary-light-dark:        #1c3238;
 $primary-inverse:           #ffffff;
 
@@ -34,7 +34,7 @@ $danger-light-dark:         #3a2434;
 $danger-inverse:            #ffffff;
 
 // Warning
-$warning:                   #f37421; // #ffc700
+$warning:                   #f37421; // metronic = #ffc700
 $warning-active:            #f1bc00;
 $warning-light:             #fff8dd;
 $warning-light-dark:        #392f28;
