+++ package.json
@@ -15,6 +15,7 @@
     "@emotion/styled": "^11.10.5",
     "@fortawesome/fontawesome-free": "^6.2.1",
     "@hookform/resolvers": "^2.9.10",
+    "@mui/icons-material": "^5.10.16",
     "@popperjs/core": "^2.11.6",
     "@types/node": "18.11.9",
     "@types/react": "18.0.25",
