+++ pages/signup.tsx
@@ -3,7 +3,7 @@ import Head from 'next/head';
 // import Skeleton from 'react-loading-skeleton';
 // import FormSignup from '../modules/signup/FormSignup';
 import Layout from '../components/layouts/Home/LayoutHome';
-import FormLogin from '../modules/login/FormLogin';
+import FormSignUp from '../modules/signup/FormSignUp';
 
 export default function SignUp(){
   return (
@@ -14,7 +14,7 @@ export default function SignUp(){
 
       {/* container-home */}
       <div className="bg-white py-9 px-4 d-flex align-items-center flex-grow-1">
-        <FormLogin />
+        <FormSignUp />
       </div>
     </Layout>
   )
